added suport for NEOPixel
During tests a bug showed up, making the sw go dead. I believe there is something to do with the NEO Pixel need to clear all the interrupts and, if a message arrives at that time, it gets lost. There is no confirmation of arrival nor a check in integrity. This should be handled diferently...